Phase 2: Structural refactor - split coordinator into bill.py, data_fetcher.py, statistics.py
