+++ custom_components/iec/iec_entity.py
@@ -2,8 +2,8 @@
 
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from custom_components.iec import IecApiCoordinator
-from custom_components.iec.commons import get_device_info, IecEntityType
+from .commons import get_device_info, IecEntityType
+from .coordinator import IecApiCoordinator
 
 
 class IecEntity(CoordinatorEntity[IecApiCoordinator]):
